+++ src/collective/solr/browser/maintenance.py
@@ -8,7 +8,6 @@ from Products.Five.browser import BrowserView
 from plone.uuid.interfaces import IUUID, IUUIDAware
 from zope.interface import implements
 from zope.component import queryUtility, queryAdapter
-from collective.indexing.indexer import getOwnIndexMethod
 from collective.solr.indexer import DefaultAdder
 from collective.solr.flare import PloneFlare
 from collective.solr.interfaces import ISolrConnectionManager
@@ -22,6 +21,17 @@ from collective.solr.parser import SolrResponse
 from collective.solr.parser import unmarshallers
 from collective.solr.utils import findObjects
 from collective.solr.utils import prepareData
+
+import pkg_resources
+
+
+try:   # pragma: no cover
+    pkg_resources.get_distribution('collective.indexing')
+    from collective.indexing.indexer import getOwnIndexMethod
+    HAS_C_INDEXING = True
+except pkg_resources.DistributionNotFound:  # pragma: no cover
+    HAS_C_INDEXING = False
+
 
 logger = getLogger('collective.solr.maintenance')
 MAX_ROWS = 1000000000
@@ -149,7 +159,8 @@ class SolrMaintenanceView(BrowserView):
         for path, obj in findObjects(self.context):
             if ICheckIndexable(obj)():
 
-                if getOwnIndexMethod(obj, 'indexObject') is not None:
+                if HAS_C_INDEXING and \
+                        getOwnIndexMethod(obj, 'indexObject') is not None:
                     log('skipping indexing of %r via private method.\n' % obj)
                     continue